Freedom in Fetters
Freedom in Fetters—a Princely Freedom.—Chopin, the last of the modern musicians, who gazed at and worshipped beauty, like Leopardi; Chopin, the Pole, the inimitable (none that came before or after him has a right to this name)—Chopin had the same princely punctilio in convention that Raphael shows in the use of the simplest traditional colours. The only difference is that Chopin applies them not to colour but to melodic and rhythmic traditions. He admitted the validity of these traditions because he was born under the sway of etiquette. But in these fetters he plays and dances as the freest and daintiest of spirits, and, be it observed, he does not spurn the chain.